+++ Artexitus.IdentityMicroservice/Artexitus.IdentityMicroservice.API/appsettings.json
@@ -12,6 +12,9 @@
     "RedisConnectionString": "redis_cache"
   },
   "Pagination": {
-    "UsersPageSize":  5
+    "UsersPageSize": 5
+  },
+  "PasswordReset": {
+    "MinutesToResetPassword":  15
   }
 }
