+++ Artexitus.IdentityMicroservice/Artexitus.IdentityMicroservice.Application/ConfigurationSections/PasswordResetSettings.cs
@@ -1,0 +1,7 @@
+﻿namespace Artexitus.IdentityMicroservice.Application.ConfigurationSections
+{
+    public record PasswordResetSettings
+    {
+        public required int MinutesToResetPassword { get; init; }
+    }
+}
