+++ Artexitus.IdentityMicroservice/Artexitus.IdentityMicroservice.Application/Extensions/ServiceCollectionExtension.cs
@@ -12,6 +12,7 @@ namespace Artexitus.IdentityMicroservice.Application.Extensions
         public static IServiceCollection AddApplication(this IServiceCollection services, IConfiguration configuration)
         {
             services.Configure<PaginationSettings>(configuration.GetSection("Pagination"));
+            services.Configure<PasswordResetSettings>(configuration.GetSection("PasswordReset"));
             services.AddAutoMapper(Assembly.GetExecutingAssembly());
             services.AddMediatR(config => 
                 config.RegisterServicesFromAssembly(typeof(ServiceCollectionExtension).Assembly)
