+++ Artexitus.IdentityMicroservice/Artexitus.IdentityMicroservice.Application/Handlers/Users/Commands/ActivateAccountHandler.cs
@@ -1,10 +1,10 @@
-﻿using Artexitus.IdentityMicroservice.Application.Interfaces;
+﻿using Artexitus.IdentityMicroservice.Domain.Repositories;
 using Artexitus.IdentityMicroservice.Contracts.Exceptions;
 using Artexitus.IdentityMicroservice.Contracts.Requests.Commands.Users;
 using MediatR;
 using Microsoft.Extensions.Logging;
 
-namespace Artexitus.IdentityMicroservice.Application.Handlers.Users
+namespace Artexitus.IdentityMicroservice.Application.Handlers.Users.Commands
 {
     public class ActivateAccountHandler : IRequestHandler<ActivateAccountCommand>
     {
