+++ Artexitus.IdentityMicroservice/Artexitus.IdentityMicroservice.Application/Handlers/Users/Commands/RefreshTokensHandler.cs
@@ -1,4 +1,4 @@
-﻿using Artexitus.IdentityMicroservice.Application.Interfaces;
+﻿using Artexitus.IdentityMicroservice.Domain.Repositories;
 using Artexitus.IdentityMicroservice.Application.Services;
 using Artexitus.IdentityMicroservice.Contracts.Exceptions;
 using Artexitus.IdentityMicroservice.Contracts.Helpers;
@@ -6,7 +6,7 @@ using Artexitus.IdentityMicroservice.Contracts.Requests.Commands.Users;
 using MediatR;
 using Microsoft.Extensions.Logging;
 
-namespace Artexitus.IdentityMicroservice.Application.Handlers.Users
+namespace Artexitus.IdentityMicroservice.Application.Handlers.Users.Commands
 {
     public class RefreshTokensHandler : IRequestHandler<RefreshTokensCommand, UserTokens>
     {
