--- Artexitus.IdentityMicroservice/Artexitus.IdentityMicroservice.Application/Interfaces/IPaginatableRepository.cs
@@ -1,10 +1,0 @@
-﻿using Artexitus.IdentityMicroservice.Domain.Entities;
-
-namespace Artexitus.IdentityMicroservice.Application.Interfaces
-{
-    public interface IPaginatableRepository<TEntity> where TEntity : class
-    {
-        Task<IPaginatedEnumerable<TEntity>> GetPaginatedAsync(int pageNumber, int pageSize, 
-            CancellationToken cancellationToken);
-    }
-}
