+++ Artexitus.IdentityMicroservice/Artexitus.IdentityMicroservice.Domain/Repositories/IUserProfileRepository.cs
@@ -1,6 +1,6 @@
 ﻿using Artexitus.IdentityMicroservice.Domain.Entities;
 
-namespace Artexitus.IdentityMicroservice.Application.Interfaces
+namespace Artexitus.IdentityMicroservice.Domain.Repositories
 {
     public interface IUserProfileRepository : IRepository<UserProfile>
     {
