+++ Artexitus.IdentityMicroservice/Artexitus.IdentityMicroservice.Infrastructure/Persistence/Repositories/UserRoleRepository.cs
@@ -1,4 +1,4 @@
-﻿using Artexitus.IdentityMicroservice.Application.Interfaces;
+﻿using Artexitus.IdentityMicroservice.Domain.Repositories;
 using Artexitus.IdentityMicroservice.Contracts.Exceptions;
 using Artexitus.IdentityMicroservice.Domain.Entities;
 using Microsoft.EntityFrameworkCore;
